Added cart functionality using sessionStorage
